Dialogs and Toasts

+++ app/favorites/favorites.component.ts
@@ -5,6 +5,8 @@ import {ListViewEventData, RadListView} from 'nativescript-telerik-ui/listview';
 import {RadListViewComponent} from 'nativescript-telerik-ui/listview/angular';
 import {ObservableArray} from 'tns-core-modules/data/observable-array';
 import {DrawerPage} from '../shared/drawer/drawer.page';
+import {confirm} from 'ui/dialogs';
+import {Toasty} from 'nativescript-toasty';
 
 @Component({
     selector: 'app-favorites',
@@ -29,8 +31,31 @@ export class FavoritesComponent extends DrawerPage implements OnInit {
     }
 
     deleteFavorite(id: number) {
-        this.favoriteservice.deleteFavorite(id)
-            .subscribe(favorites => this.favorites = new ObservableArray<Dish>(favorites), errmess => this.errMess = errmess);
+        console.log('delete ', id);
+
+        let options = {
+            title: "Confirm Delete",
+            message: 'Do you want to delete Dish ' + id,
+            okButtonText: "Yes",
+            cancelButtonText: "No",
+            neutralButtonText: "Cancel"
+        };
+
+        confirm(options)
+            .then((result: boolean) => {
+                if (result) {
+                    this.favorites = null;
+
+                    this.favoriteservice.deleteFavorite(id)
+                        .subscribe(favorites => {
+                            const toast = new Toasty("Deleted Dish " + id, "short", "bottom");
+                            toast.show();
+                            this.favorites = new ObservableArray<Dish>(favorites);
+                        }, errmess => this.errMess = errmess);
+                } else {
+                    console.log('Delete cancelled');
+                }
+            });
     }
 
     public onCellSwiping(args: ListViewEventData) {
@@ -51,7 +76,7 @@ export class FavoritesComponent extends DrawerPage implements OnInit {
         var rightItem = swipeView.getViewById('delete-view');
         swipeLimits.left = leftItem.getMeasuredWidth();
         swipeLimits.right = rightItem.getMeasuredWidth();
-        swipeLimits.threshold = leftItem.getMeasuredWidth()/2;
+        swipeLimits.threshold = leftItem.getMeasuredWidth() / 2;
     }
 
     public onSwipeCellFinished(args: ListViewEventData) {
